增加了简陋的缓存功能
到头来还是在写屎山，就和我的人生一样

+++ src/main/kotlin/github/samyycx/miraisetu/util/DownloadUtil.kt
@@ -1,5 +1,6 @@
 package github.samyycx.miraisetu.util
 
+import github.samyycx.miraisetu.MiraiSetu
 import org.apache.commons.io.FileUtils
 import java.io.File
 import java.io.IOException
@@ -7,10 +8,15 @@ import java.net.URL
 
 object DownloadUtil {
 
-    fun downloadImage(url: String): File? {
-        val ext = url.split(".").last()
-        val file = File("./cache/temp.${ext}")
-        println("2")
+    fun downloadImage(url: String, folder: File): File? {
+
+        val localFile = File(url)
+        if (localFile.exists()) {
+            return localFile
+        }
+
+        val fileName = url.split("/")
+        val file = File(folder, fileName.last())
         try {
             FileUtils.copyURLToFile(
                 URL(url),
